Bug Patch.
Resolved a problem in the algorithm that was being used to detect API routes within an API description.
The conditional statement was incorrect, and it was allowing any object with an ID field to be detected as an API route.
Instead, the conditional now only allows objects with ID, method and uri fields, or rel, method, and uri fields.